fix cast error and set batch to 205 (default 2048)

+++ src/main/java/be/horafrost/esb/ArticleEnricher.java
@@ -19,7 +19,7 @@ public class ArticleEnricher implements AggregationStrategy {
 		
 		Map<String, Object> responseMap = response.get(0);
 		
-		if((Integer)responseMap.get("category") != 29) 
+		if((Long)responseMap.get("category") != 29) 
 			return oldExchange;
 		
 		originalBody.code = ((String)responseMap.get("code"));
